Consolidate web controllers under \Rogue\Http\Controllers\Web.
