+++ app/Http/Controllers/Web/CampaignIdsController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Legacy\Web;
+namespace Rogue\Http\Controllers\Web;
 
 use Rogue\Types\Cause;
 use Rogue\Models\Campaign;
