+++ app/Http/Controllers/Web/UsersController.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Rogue\Http\Controllers\Legacy\Web;
+namespace Rogue\Http\Controllers\Web;
 
 use Rogue\Models\User;
 use Illuminate\Http\Request;
